% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afcd2}
\alias{afcd2}
\title{Aquatic Foods Composition Database (AFCD) - w/out scientific names}
\format{
A data frame with the following attributes:
\describe{
  \item{taxa_name_source}{English or original food name}
  \item{taxa_name}{Food name}
  \item{study_type}{Study type (peer-reviewed paper or FCT report)}
  \item{study_id}{Study id}
  \item{iso3}{ISO3 of source country(s)}
  \item{country}{Source country(s)}
  \item{prod_catg}{Production category (farmed, wild capture, unknown)}
  \item{food_part}{Part of food}
  \item{food_prep}{Preparation of food}
  \item{food_name}{Name of food, in English}
  \item{food_name_orig}{Name of food, in original data}
  \item{fct_code_orig}{FCT code, in original data}
  \item{food_id}{Food id}
  \item{edible_prop}{Edible proportion}
  \item{notes}{Notes}
  \item{nutrient_type}{Nutient type}
  \item{nutrient}{Nutient name}
  \item{nutrient}{Nutient name, in the raw AFCD}
  \item{nutrient_desc}{Description of nutrient}
  \item{nutrient_units}{Units per 100g of edible food portion}
  \item{value}{Value}
}
}
\source{
Golden CD, Koehn JZ, Shepon A, Passarelli S, Free CM, Viana DF, Matthey H, Eurich JG, Gephart JA, Fluet-Chouinnard E, Nyboer EA, Lynch AJ, Kjellevold M, Bromage S, Charlebois P, Barange M, Vannuccini S, Cao L, Kleisner KM, Rimm EB, Danaei G, DeSisto C, Kelahan H, Fiorella KJ, Little DC, Allison EH, Fanzo J, Thilsted SH (2021) Aquatic foods to nourish nations. Nature 598: 315-320.
}
\usage{
afcd2
}
\description{
A cleaned version of the Aquatic Foods Composition Database (AFCD). This dataset includes observations with foods identified by general names (not scientific names).
}
\keyword{datasets}
