% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afcd_parts}
\alias{afcd_parts}
\title{Aquatic Foods Composition Database (AFCD) food parts key}
\format{
A data frame with the following attributes:
\describe{
  \item{food_part}{Food part type}
  \item{n}{Number of observations}
}
}
\source{
Golden CD, Koehn JZ, Shepon A, Passarelli S, Free CM, Viana DF, Matthey H, Eurich JG, Gephart JA, Fluet-Chouinnard E, Nyboer EA, Lynch AJ, Kjellevold M, Bromage S, Charlebois P, Barange M, Vannuccini S, Cao L, Kleisner KM, Rimm EB, Danaei G, DeSisto C, Kelahan H, Fiorella KJ, Little DC, Allison EH, Fanzo J, Thilsted SH (2021) Aquatic foods to nourish nations. Nature 598: 315-320.
}
\usage{
afcd_parts
}
\description{
Parts key for the Aquatic Foods Composition Database (AFCD).
}
\keyword{datasets}
