% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afcd_refs}
\alias{afcd_refs}
\title{Aquatic Foods Composition Database (AFCD) reference key}
\format{
A data frame with the following attributes::
\describe{
  \item{study_type}{Study type (Peer-reviewed literature or FCT table)}
  \item{study_id}{Study id}
  \item{citation}{Citation}
  \item{doi}{DOI for peer-reviewed sources and dataset link for FCT sources}
  \item{region}{Study region (peer-reviewed only)}
  \item{database}{Nuutrient database (FCT sources only)}
  \item{units}{Nutrients per unit (FCT sources only)}
}
}
\source{
Golden CD, Koehn JZ, Shepon A, Passarelli S, Free CM, Viana DF, Matthey H, Eurich JG, Gephart JA, Fluet-Chouinnard E, Nyboer EA, Lynch AJ, Kjellevold M, Bromage S, Charlebois P, Barange M, Vannuccini S, Cao L, Kleisner KM, Rimm EB, Danaei G, DeSisto C, Kelahan H, Fiorella KJ, Little DC, Allison EH, Fanzo J, Thilsted SH (2021) Aquatic foods to nourish nations. Nature 598: 315-320.
}
\usage{
afcd_refs
}
\description{
Reference key for the Aquatic Foods Composition Database (AFCD). The AFCD utilizes a mixture of values from Food Composition Tables (FCTs) and from the peer-reviewed literature.
}
\keyword{datasets}
