% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_nutrients.R
\name{species_nutrients}
\alias{species_nutrients}
\title{Match nutritional content to nearest taxa}
\usage{
species_nutrients(sci_name, prep, part, nut)
}
\arguments{
\item{sci_name}{A character vector of species scientific names}

\item{prep}{A character vector of fish preparation types}

\item{part}{A character vector of fish parts}

\item{nut}{A character vector of nutrients to search for}
}
\value{
A dataframe containing species, nutrient, nutrient units, nutritional value (value), closest taxa match (taxa_match), number of observations which nutritional value was extracted (count), standard error (se), startard deviation (ssd) and confidence intervals
}
\description{
Get nutritional information for aquatic animals from the Aquatic Foods Composition Database. This function uses the hierarchical approach described in Golden et al 2021, where nutritional content is assigned based on the closest taxa possible.
}
\examples{
# Get nutritional info
# species_nutrients(sci_name = c("Homarus americanus", "Pangasius pangasius"), prep = "raw", part = "muscle tissue", nut = c("Iron, total", "Zinc"))
}
