% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_table}
\alias{taxa_table}
\title{Comprehensive taxonomic table for aquatic species}
\format{
A data frame with the following attributes:
\describe{
  \item{genus}{species Genus}
  \item{family}{species Family}
  \item{order}{species Order}
  \item{class}{species Class}
  \item{phylum}{species Phylum}
  \item{kigdom}{species Kingdom}
}
}
\source{
Golden CD, Koehn JZ, Shepon A, Passarelli S, Free CM, Viana DF, Matthey H, Eurich JG, Gephart JA, Fluet-Chouinnard E, Nyboer EA, Lynch AJ, Kjellevold M, Bromage S, Charlebois P, Barange M, Vannuccini S, Cao L, Kleisner KM, Rimm EB, Danaei G, DeSisto C, Kelahan H, Fiorella KJ, Little DC, Allison EH, Fanzo J, Thilsted SH (2021) Aquatic foods to nourish nations. Nature 598: 315-320.
}
\usage{
taxa_table
}
\description{
Taxonomic information for all species in AFCD. Sources: FishBase, SeaLifeBase and WoRMS.
}
\keyword{datasets}
