% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afcd}
\alias{afcd}
\title{Aquatic Foods Composition Database (AFCD)}
\format{
A data frame with the following attributes:
\describe{
  \item{sciname}{Scientific name}
  \item{study_id}{Study id}
  \item{iso3}{ISO3 of source country(s)}
  \item{fao3}{}
  \item{prod_catg}{Production category (farmed, wild capture, unknown)}
  \item{food_part}{Part of food}
  \item{food_prep}{Preparation of food}
  \item{food_name}{Name of food, in English}
  \item{food_name_orig}{Name of food, in original data}
  \item{notes}{Notes}
  \item{nutrient}{Nuteient}
  \item{description}{Description of nutrient}
  \item{units}{Units per 100g of edible food portion}
  \item{value}{Value}
}
}
\source{
Golden CD, Koehn JZ, Shepon A, Passarelli S, Free CM, Viana DF, Matthey H, Eurich JG, Gephart JA, Fluet-Chouinnard E, Nyboer EA, Lynch AJ, Kjellevold M, Bromage S, Charlebois P, Barange M, Vannuccini S, Cao L, Kleisner KM, Rimm EB, Danaei G, DeSisto C, Kelahan H, Fiorella KJ, Little DC, Allison EH, Fanzo J, Thilsted SH (2021) Aquatic foods to nourish nations. Nature 598: 315-320.
}
\usage{
afcd
}
\description{
A cleaned version of the Aquatic Foods Composition Database (AFCD).
}
\keyword{datasets}
